fetch function extracted to a custom hook
