+++ src/pages/Dashboard/components/AccountPreferences/AccountPreferences.jsx
@@ -1,51 +1,13 @@
-import React, { useState, useCallback, useEffect } from "react"
+import React, { useState } from "react"
 import styles from "./AccountPreferences.module.css"
 import Button from "../Button/Button"
-import { API_LINK } from "../../../../utils/api"
-import { toast } from "react-hot-toast"
 import Loader from "../../../../Components/Loader/Loader"
+import useFetchUserInfo from "../../../../hooks/useFetchUserInfo"
 
 const AccountPreferences = () => {
   const [enableInput, setEnableInput] = useState(true)
-  const [loading, setLoading] = useState(false)
 
-  const [fetchedUserData, setFetchedUserData] = useState({})
-
-  const fetchUserHandler = useCallback(async () => {
-    setLoading(true)
-    try {
-      const token = sessionStorage.getItem("token")
-      const getUserData = await fetch(`${API_LINK}/api/profile/user`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${token}`
-        }
-      })
-      const userData = await getUserData.json()
-      // console.log(getUserData)
-      if (!getUserData.ok || !getUserData) {
-        setLoading(false)
-        toast.error(`${userData?.msg} Please login again!`, {
-          duration: 4000,
-          position: "top-center",
-          // Styling
-          style: { fontSize: "13px" },
-          className: ""
-        })
-      }
-      // console.log(userData)
-      setFetchedUserData(userData.user)
-      setLoading(false)
-    } catch (error) {
-      console.log(error)
-      setLoading(false)
-    }
-  }, [])
-
-  useEffect(() => {
-    fetchUserHandler()
-  }, [fetchUserHandler])
+  const { loading, fetchedUserData } = useFetchUserInfo(`api/profile/user`)
 
   return (
     <div className={styles.accountPreferences}>
